Put `fixed_node` macros into separate module (#9)
